Added function getWindowHandle() for JCEF support. (Fixed)

+++ skiko/src/jvmMain/kotlin/org/jetbrains/skiko/Drawable.kt
@@ -4,6 +4,6 @@ interface Drawable {
     fun redrawLayer()
     fun updateLayer()
     fun disposeLayer()
-    val windowHandle: Int
+    val windowHandle: Long
     val contentScale: Float
 }
