+++ skiko/src/jvmMain/kotlin/org/jetbrains/skiko/HardwareLayer.kt
@@ -15,7 +15,7 @@ open class HardwareLayer : Canvas(), Drawable {
     external override fun redrawLayer()
     external override fun updateLayer()
     external override fun disposeLayer()
-    override val windowHandle: Int
+    override val windowHandle: Long
         external get
     override val contentScale: Float
         external get
